// src/tournament-management/types/enums.ts
export enum TournamentFormat {
  LEAGUE = 'LEAGUE',
  GROUP_KNOCKOUT = 'GROUP_KNOCKOUT',
  KNOCKOUT = 'KNOCKOUT'
}

export enum MatchFormat {
  SINGLE_MATCH = 'SINGLE_MATCH',
  HOME_AWAY = 'HOME_AWAY'
}

export enum KnockoutQualification {
  POINTS = 'POINTS',
  GROUP_POSITION = 'GROUP_POSITION'
}

export enum Tiebreaker {
  GOAL_DIFFERENCE = 'GOAL_DIFFERENCE',
  HEAD_TO_HEAD = 'HEAD_TO_HEAD',
  GOALS_SCORED = 'GOALS_SCORED'
}

export enum TournamentStatus {
  NOT_STARTED = 'NOT_STARTED',
  IN_PROGRESS = 'IN_PROGRESS',
  COMPLETED = 'COMPLETED'
} 